feat: move catalog source build to a package (#251)
