+++ catalog-build/package/catalog_build/__init__.py
@@ -1,0 +1,3 @@
+from . import build
+
+build_files = build.build_files
