+++ catalog-build/package/setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+setup(
+	name="catalog_build",
+	version="1.0.0",
+	packages=["catalog_build"],
+	install_requires=["pandas", "requests", "PyYAML"],
+)
